+++ Source/AnimusExMachina/Private/StateTree/Tasks/AxMGlobalTask_Config.cpp
@@ -28,6 +28,8 @@ EStateTreeRunStatus FAxMGlobalTask_Config::EnterState(
 
 	// Copy all values from the Data Asset to bindable outputs
 	InstanceData.EngagementRange = ConfigAsset->EngagementRange;
+	InstanceData.LeashRadius = ConfigAsset->LeashRadius;
+	InstanceData.PatrolWaitDuration = ConfigAsset->PatrolWaitDuration;
 	InstanceData.PatrolMovementSpeed = ConfigAsset->PatrolMovementSpeed;
 	InstanceData.InvestigateMovementSpeed = ConfigAsset->InvestigateMovementSpeed;
 	InstanceData.CombatMovementSpeed = ConfigAsset->CombatMovementSpeed;